moved combine.py inside pipeline/
